pyproject.toml
[project]
name = "nitter-timeline"
version = "0.1.0"
description = "Web UI that aggregates Nitter RSS feeds into a Twitter-like timeline"
authors = [
	{ name = "Your Name", email = "you@example.com" }
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
	"fastapi>=0.111.0,<1.0.0",
	"uvicorn[standard]>=0.30.0,<1.0.0",
	"httpx>=0.27.0,<1.0.0",
	"feedparser>=6.0.11,<7.0.0",
	"jinja2>=3.1.4,<4.0.0",
	"python-dateutil>=2.9.0.post0,<3.0.0",
	"cachetools>=5.3.3,<6.0.0",
	"pydantic>=1.10.15,<2.0.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.23", "ruff>=0.5.5", "mypy>=1.11.0", "types-python-dateutil"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
